+++ src/scripts/cache/cache.ts
@@ -1,3 +1,5 @@
+import type {SAFE_ANY} from '@helpers/type';
+
 import {existsSync, mkdirSync, readFileSync, unlinkSync, writeFileSync} from 'node:fs';
 
 import {oraExecCmd} from '../helpers';
@@ -12,6 +14,7 @@ export interface CacheData {
     formatDate: string;
     expiredDate: number;
     expiredFormatDate: string;
+    execResult: SAFE_ANY;
   };
 }
 
@@ -37,7 +40,8 @@ export function getCacheData(): CacheData {
 export function cacheData(
   packageName: string,
   packageData: {
-    version: string;
+    version?: string;
+    execResult?: SAFE_ANY;
   },
   cacheData?: CacheData
 ) {
@@ -48,7 +52,7 @@ export function cacheData(
   const expiredDate = +now + cacheTTL;
 
   data[packageName] = {
-    ...packageData,
+    ...(packageData as SAFE_ANY),
     date: now,
     expiredDate,
     expiredFormatDate: new Date(expiredDate).toString(),
@@ -70,17 +74,13 @@ function ttl(n: number) {
   return now() - n;
 }
 
-function isExpired(packageName: string, cacheData?: CacheData) {
+export function isExpired(packageName: string, cacheData?: CacheData) {
   const data = cacheData ?? getCacheData();
   const pkgData = data[packageName];
 
   if (!pkgData?.expiredDate) return true;
 
-  if (ttl(pkgData.expiredDate) > 0) {
-    return true;
-  }
-
-  return false;
+  return ttl(pkgData.expiredDate) > 0;
 }
 
 export async function getPackageData(packageName: string) {
@@ -90,12 +90,26 @@ export async function getPackageData(packageName: string) {
   // If expired or don't exist then init data
   if (isExpiredPkg) {
     const version = await oraExecCmd(
-      `Fetching ${packageName} latest version`,
-      `npm view ${packageName} version`
+      `npm view ${packageName} version`,
+      `Fetching ${packageName} latest version`
     );
 
     cacheData(packageName, {version}, data);
   }
 
   return data[packageName]!;
+}
+
+export async function getCacheExecData(key: string) {
+  const data = getCacheData();
+  const isExpiredPkg = isExpired(key, data);
+
+  // If expired or don't exist then init data
+  if (isExpiredPkg) {
+    const execResult = await oraExecCmd(key);
+
+    cacheData(key, {execResult}, data);
+  }
+
+  return data[key]!.execResult;
 }